+++ packages/ui-components/docs/changelog.md
@@ -2,6 +2,10 @@
 title: Changelog
 ---
 
+## Next
+
+-   Memoize `Table` rows to improve performance.
+
 ## 1.0.0-a.1
 
 -   Initial release